refactor: 테스트 패키지 util 에서 step1로 변경
